+++ tests/testthat/test-init.R
@@ -262,3 +262,10 @@ test_that("init() respects user-requested snapshot type", {
   expect_false(renv_package_installed("toast"))
   expect_equal(settings$snapshot.type(), "explicit")
 })
+
+test_that("init() respected Remotes in a project DESCRIPTION file", {
+  project <- renv_tests_scope("skeleton")
+  writeLines("Depends: skeleton\nRemotes: kevinushey/skeleton", con = "DESCRIPTION")
+  init()
+  expect_true(renv_package_installed("skeleton"))
+})